requirements.txt
django-autoslug==1.9.8
graphene==2.1.8
graphene-django==2.13.0
django-filter==2.3.0
django-webpack-loader==0.7.0
django-cors-headers==3.5.0
django-ckeditor==6.0.0
